backend/pyproject.toml
[tool.poetry]
name = "fitmentor"
version = "0.1.0"
description = "AI Fitness Coach backend"
authors = ["AK <anushilswift05@gmail.com>"]
readme = "README.md"
packages = [{include = "fitmentor", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
structlog = "^24.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"
tenacity = "^9.0.0"
redis = "^5.1.0"
aioboto3 = "^13.2.0"
portkey-ai = "^1.10.0"
openai = "^1.54.0"
apscheduler = "^3.10.4"
jinja2 = "^3.1.4"
weasyprint = "^62.3"
greenlet = "^3.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.7.0"
black = "^24.10.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
